requirements.txt
appnope==0.1.3
asgiref==3.3.2
asttokens==2.2.1
autopep8==1.7.0
backcall==0.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.3
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==3.2.18
django-allauth==0.54.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
executing==1.2.0
factory-boy==3.2.1
Faker==18.11.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.7.0
pluggy==1.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.15.1
PyJWT==1.7.1
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-warnings==0.3.1
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
pyzmq==25.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
toml==0.10.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.6
